cv-forge/requirements.txt
python-docx>=1.0.0
PyPDF2>=3.0.1
pdfplumber>=0.10.0
reportlab>=4.0.9
Pillow>=10.1.0
requests>=2.31.0
